requirements.txt
fastapi~=0.128.0
xgboost==3.0.0
pydicom~=3.0.1
numpy~=2.1.3
matplotlib~=3.10.1
pillow~=11.1.0
tqdm~=4.67.1
opencv-python~=4.11.0.86
SimpleITK~=2.4.1
joblib~=1.4.2
dask-ml==2025.1.0
tensorflow==2.19.0
pandas~=2.2.3
scikit-image~=0.25.2
scipy~=1.15.2
waitress~=3.0.2